refactor: move utils and config to parent folder
